# [iOS GPU][Perf][1/n] Use aten::contiguous instead of permuting weights manually