--- a/aten/src/ATen/native/metal/MetalPrepackOpContext.h
+++ b/aten/src/ATen/native/metal/MetalPrepackOpContext.h
@@ -68,30 +68,6 @@ class Conv2dOpContext : public torch::jit::CustomClassHolder {
   std::function<void(void*)> releaseCallback = nullptr;
 };
 
-// The MPSCNNConvolution class takes weights in the order
-// [outputChannels][kernelHeight][kernelWidth][inputChannels/groups].
-static inline std::vector<float> permuteWeights(
-    const float* src,
-    const std::vector<int64_t>& sizes) {
-  const int64_t M = sizes[0];
-  const int64_t Cf = sizes[1];
-  const int64_t kH = sizes[2];
-  const int64_t kW = sizes[3];
-  std::vector<float> packedWeights(M * kH * kW * Cf);
-  for (auto m = 0; m < M; ++m) {
-    for (auto c = 0; c < Cf; ++c) {
-      for (auto kh = 0; kh < kH; ++kh) {
-        for (auto kw = 0; kw < kW; ++kw) {
-          int64_t oc = m * kH * kW * Cf + kh * kW * Cf + kw * Cf + c;
-          int64_t ic = m * Cf * kH * kW + c * kH * kW + kh * kW + kw;
-          packedWeights[oc] = src[ic];
-        }
-      }
-    }
-  }
-  return packedWeights;
-}
-
 } // namespace metal
 } // namespace native
 } // namespace at