--- a/aten/src/ATen/native/metal/MetalUtils.h
+++ b/aten/src/ATen/native/metal/MetalUtils.h
@@ -25,6 +25,30 @@ std::vector<float> NC4ToNCHW(
     const float* src,
     const std::vector<int64_t>& sizes);
 
+// The MPSCNNConvolution class takes weights in the order
+// [outputChannels][kernelHeight][kernelWidth][inputChannels/groups].
+static inline std::vector<float> permuteWeights(
+    const float* src,
+    const std::vector<int64_t>& sizes) {
+  const int64_t M = sizes[0];
+  const int64_t Cf = sizes[1];
+  const int64_t kH = sizes[2];
+  const int64_t kW = sizes[3];
+  std::vector<float> packedWeights(M * kH * kW * Cf);
+  for (auto m = 0; m < M; ++m) {
+    for (auto c = 0; c < Cf; ++c) {
+      for (auto kh = 0; kh < kH; ++kh) {
+        for (auto kw = 0; kw < kW; ++kw) {
+          int64_t oc = m * kH * kW * Cf + kh * kW * Cf + kw * Cf + c;
+          int64_t ic = m * Cf * kH * kW + c * kH * kW + kh * kW + kw;
+          packedWeights[oc] = src[ic];
+        }
+      }
+    }
+  }
+  return packedWeights;
+}
+
 // When copying the result back to a CPU tensor, the memory format becomes NCHW.
 // Thus,we compute the strides based on contiguous memory format.
 static inline std::vector<int64_t> compute_strides(