Stack from [ghstack](https://github.com/ezyang/ghstack):
* #56076 [iOS GPU][MaskRCNN][Not For Landing] Run the maskrcnn model in playground
* #56075 [iOS GPU][Kernel][MaskRCNN] Implement RoIAlign in Metal shaders using Sampler
* #57668 [iOS GPU][perf][5/n] Replace std::vector with  IntArrayRef and SmallVector
* #57667 [iOS GPU][Perf][4/n] Reuse the same command buffer when copying results to CPU
* #57666 [iOS GPU][Perf][3/n] Cache the compuation pipeline state object
* #57665 [iOS GPU][Perf][2/n] Prepack linear + Fuse relu/hardtanh
* **#57664 [iOS GPU][Perf][1/n] Use aten::contiguous instead of permuting weights manually**

Manually permuting weights is slower than using aten::contiguous. This will improve the model loading time at runtime especially on low-end devices. Some numbers from the Unet model. Average 6x faster.

- iPhone 12
    - before - 26.252 ms
    - after - 4.727 ms
- iPhone 11
   - before - 29.638 ms
   - after - 5.012 ms
- iPhone X
     - before - 33.257 ms
     - after - 5.481 ms
- iPhone 8
     - before - 33.335 ms
     - after - 5.83 ms
- iPhone 7
     - before - 36.144 ms
     - after - 6.232 ms
- iPhone 6s
     - before - 47.977 ms
     - after - 6.998 ms

Differential Revision: [D28087911](https://our.internmc.facebook.com/intern/diff/D28087911/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D28087911/)!